Patched s3Client.destroy() in move-object.js

+++ nodes/move-object/move-object.js
@@ -121,7 +121,7 @@ function inputHandler(n, RED) {
       msg.error = err;
       send(msg);
     } finally {
-      s3Client.destroy();
+      if(s3Client) s3Client.destroy();
       /* Dereference vars */
       s3Client = null;
       /*********************/
